rag-climate-action-tracker/requirements.txt
# Web Scraping & Data Extraction
scrapy==2.12.0
selenium==4.28.1
pdfplumber==0.10.3
beautifulsoup4==4.12.3
requests==2.31.0

# Data Analysis & Scientific Computing
numpy==1.26.4
pandas==2.2.3
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.3
seaborn==0.13.2

# Natural Language Processing
nltk==3.8.1
spacy==3.7.4
gensim==4.3.2
sentence-transformers==2.6.0
huggingface_hub==0.21.4
markdown==3.5.2

# Database & Storage
SQLAlchemy==2.0.30
pgvector==0.2.1
psycopg2-binary==2.9.9

# Development & Debugging
ipykernel==6.21.2
jupyterlab==3.6.1
ipdb==0.13.13
python-dotenv==1.0.1
